refactor: views structure refactor
