+++ src/App.tsx
@@ -3,15 +3,15 @@ import "./App.css";
 import styled from "styled-components/macro";
 import { Switch, Route, Redirect } from "react-router-dom";
 import { useSpring, animated } from "react-spring";
-import Biography from "./Bio/Bio";
-import Portfolio from "./Portfolio/Portfolio";
-import Home from "./Home/Home";
+import Biography from "./Views/Bio/Bio";
+import Portfolio from "./Views/Portfolio/Portfolio";
+import Home from "./Views/Home/Home";
 import { Top } from "./Components/Misc/Misc";
 import { Link } from "react-router-dom";
 import Spinner from "./Components/Spinner/Spinner";
 import { Container } from "@material-ui/core";
 import Footer from "./TopStyle/Footer";
-import { Inspire } from "./Inspire/Inspire";
+import { Inspire } from "./Views/Inspire/Inspire";
 import breakpoints from "./Components/Common/breakpoints";
 import { Header } from "./Components/Header/Header";
 
@@ -65,7 +65,6 @@ export default App;
 
 const HomeCont = styled(animated.div)`
   display: flex;
-  // min-height: 100vh;
   background-color: ${({ theme }) => theme.colours.newBlack};
 `;
