+++ src/Views/Portfolio/AnimatedPortfolio.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import styled from "styled-components";
 import P5Wrapper from "react-p5-wrapper";
-import sketch from "../animation/drawMatrix";
+import sketch from "../../animation/drawMatrix";
 
 export default function AnimatedPortfolio(): JSX.Element {
   const [dimensions, setDimensions] = useState({ height: 200, width: 500 });
