+++ src/Views/Portfolio/LegacyPortfolio.tsx
@@ -1,7 +1,7 @@
 import React, { useMemo, useState } from "react";
 import styled from "styled-components";
-import Card from "../Components/GalleryCard/GalleryCard";
-import data from "../Data/data.json";
+import Card from "../../Components/GalleryCard/GalleryCard";
+import data from "../../Data/data.json";
 import { Pagination } from "@mui/material";
 
 export default function LegacyPortfolio(): JSX.Element {
